JDK-8201539: Fix DirectWrite crash on Monocle.
COM (aka OLE) is initialized in two places, at the creation of a
WinApplication and in directwrite.cpp when IWICImagingFactory is
created for the first time.

The IWICImagingFactory initialization in directwrite.cpp initializes
COM, creates an IWICImagingFactory instance, and then uninitializes COM.

When Monocle's HeadlessPlatform is used calling the CreateBitmap method
of IWICImagingFactory causes a crash. Without Monocle, on the normal
WinPlatform, calling CreateBitmap does *not* cause a crash. In both
cases COM should be un-initialized by the IWICImagingFactory
initialization method.

What accounts for the difference?